Create the post viewing logic.
This functionality allows non-registered users from viewing any blog
through a subdomain. `www` should be reserved for the main website.
To do this, create a new `view-blog` template and JavaScript file. Also,
change the Express.js configuration to use three routers which are
configured to route to the `www` subdomain (for the main site), route to
all subdomains that aren't `www` (for blogs), and route anything (for
the public API and hosted files).